Move the copr Makefile to the top-level directory
This is also not what I expected to happen, but on the bright side it
will save me the trouble of making per-package symlinks.

Move Makefile.copr to .copr/Makefile, and modify the srpm target to
handle being run from an arbitrary directory that may or may not have a
spec file in it. Remove python-nss/.copr/Makefile since I guess I don't
need it.