Update to python-nss-1.0.1^20210803hg9de14a6f77e2-6
- Correct the license files

wheel and setuptools disagree on the format of the `license-files`
metadata. Newer versions of wheel use the data from setuptools instead
of parsing it itself, which was causing duplicated license files in the
final package. Instead, remove the license-files information from
setup.cfg and let the automatic pattern install the licenses to
dist-info, and remove the explicit license file list in %files.